requirements.txt
Flask==3.1.0
opencv-python==4.10.0.84
face-recognition==1.3.0
numpy==2.1.1
Pillow==10.4.0
